+++ packages/core/src/node.ts
@@ -6,10 +6,14 @@ export type {
   GetImgParams,
   GetImgSourceArgs,
   GetImgSource,
+  GetSharpPipelineArgs,
+  GetSharpPipeline,
   ImgParams,
   ImgSource,
+  SharpConfig,
 } from "./utils";
 
 export { getImgResponse } from "./node/index";
 export { getImgPlaceholder } from "./placeholder";
 export { getImgMetadata, type Metadata } from "./meta";
+export { getDefaultSharpPipeline } from "./utils";